Update angular_corr.js
Min and max mixing for a2, a4 plots were read assuming integers, when they should be assumed to be floats.

+++ js/angular_corr.js
@@ -660,8 +660,8 @@ function updateMixingSamples(){
 }
 
 function updateMixLimits(){
-    dataStore.minMix = parseInt(document.getElementById('minMix').value,10);
-    dataStore.maxMix = parseInt(document.getElementById('maxMix').value,10);
+    dataStore.minMix = parseFloat(document.getElementById('minMix').value,10);
+    dataStore.maxMix = parseFloat(document.getElementById('maxMix').value,10);
     recalculate();
 
 }
